ministerio con tests arreglados

+++ src/ar/edu/unq/po2/RepasoParcial/Ministerio/Empleado.java
@@ -5,5 +5,5 @@ public interface Empleado {
 
 	public String rol();
 
-	public float sueldoDiario();
+	public int sueldoDiario();
 }
